Specify Organization indexed on search service (#356)
* Specify Organization indexed on search service

a null indexedEntityId specifies it is the users entire vault.
otherwise, organizations specify their id to signify the index is a subset.

user's vault will re-index if the indexed entity does not match the
users id or null. at the moment, user's vault does not set userId
because indexing occurs in the setter for decryptedCipherCache,
which cannot be asynchronous

* Linter fix

(cherry picked from commit 090ad790f5ace3526ce8202d2ac74751e6618587)